+++ EliteChattSystem/src/Client/ChatClient.java
@@ -1,5 +1,4 @@
 package Client;
-
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.BufferedReader;
@@ -14,64 +13,34 @@ import javax.swing.JScrollPane;
 import javax.swing.JTextArea;
 import javax.swing.JTextField;
 
-/**
- * A simple Swing-based client for the chat server.  Graphically
- * it is a frame with a text field for entering messages and a
- * textarea to see the whole dialog.
- *
- * The client follows the Chat Protocol which is as follows.
- * When the server sends "SUBMITNAME" the client replies with the
- * desired screen name.  The server will keep sending "SUBMITNAME"
- * requests as long as the client submits screen names that are
- * already in use.  When the server sends a line beginning
- * with "NAMEACCEPTED" the client is now allowed to start
- * sending the server arbitrary strings to be broadcast to all
- * chatters connected to the server.  When the server sends a
- * line beginning with "MESSAGE " then all characters following
- * this string should be displayed in its message area.
- */
 public class ChatClient {
-
+	//Clientens Reader Skrivare JFrame Textfield o TextArea 
     BufferedReader in;
     PrintWriter out;
     JFrame frame = new JFrame("Chatter");
     JTextField textField = new JTextField(40);
     JTextArea messageArea = new JTextArea(8, 40);
 
-    /**
-     * Constructs the client by laying out the GUI and registering a
-     * listener with the textfield so that pressing Return in the
-     * listener sends the textfield contents to the server.  Note
-     * however that the textfield is initially NOT editable, and
-     * only becomes editable AFTER the client receives the NAMEACCEPTED
-     * message from the server.
-     */
+    
     public ChatClient() {
-
-        // Layout GUI
-        textField.setEditable(false);
+    	
+    	textField.setEditable(false);
         messageArea.setEditable(false);
-        frame.getContentPane().add(textField, "North");
+        //Frame layout
+        frame.getContentPane().add(textField, "South");
         frame.getContentPane().add(new JScrollPane(messageArea), "Center");
         frame.pack();
-
-        // Add Listeners
+        //textField actionlistern ifall de finns text i textfield så skriver den ut sedan sätter texten t "" aka tom
         textField.addActionListener(new ActionListener() {
-            /**
-             * Responds to pressing the enter key in the textfield by sending
-             * the contents of the text field to the server.    Then clear
-             * the text area in preparation for the next message.
-             */
             public void actionPerformed(ActionEvent e) {
-                out.println(textField.getText());
-                textField.setText("");
+            	if(!textField.getText().equalsIgnoreCase("")) {
+            		out.println(textField.getText());
+            		textField.setText("");
+            	}
             }
         });
     }
-
-    /**
-     * Prompt for and return the address of the server.
-     */
+    //När man startar programmet kmr en JOptionPane ruta där man skriver in IP address aka lokal aka 127.0.0.1
     private String getServerAddress() {
         return JOptionPane.showInputDialog(
             frame,
@@ -80,9 +49,7 @@ public class ChatClient {
             JOptionPane.QUESTION_MESSAGE);
     }
 
-    /**
-     * Prompt for and return the desired screen name.
-     */
+    //Samma ruta som innan fast man ska skriva in vad man nickkar in-game 
     private String getName() {
         return JOptionPane.showInputDialog(
             frame,
@@ -90,13 +57,9 @@ public class ChatClient {
             "Screen name selection",
             JOptionPane.PLAIN_MESSAGE);
     }
-
-    /**
-     * Connects to the server then enters the processing loop.
-     */
+    //Connectar till servern efter man  skrivit in IP och namn 
     private void run() throws IOException {
 
-        // Make connection and initialize streams
         String serverAddress = getServerAddress();
         Socket socket = new Socket(serverAddress, 9001);
         in = new BufferedReader(new InputStreamReader(
