+++ EliteChattSystem/src/Server/Handler.java
@@ -1,0 +1,67 @@
+package Server;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.PrintWriter;
+import java.net.Socket;
+
+	public class Handler extends Thread {
+        private String name;
+        private Socket socket;
+        private BufferedReader in;
+        private PrintWriter out;
+        private ChatServer cs = new ChatServer();
+
+        public Handler(Socket socket) {
+            this.socket = socket;
+        }
+        public void run() {
+            try {
+
+                in = new BufferedReader(new InputStreamReader(socket.getInputStream()));
+                out = new PrintWriter(socket.getOutputStream(), true);
+                
+                while (true) {
+                    out.println("SUBMITNAME");
+                    name = in.readLine();
+                    if (name == null) {
+                        return;
+                    }
+                    synchronized (cs.names) {
+                        if (!cs.names.contains(name)) {
+                            cs.names.add(name);
+                            break;
+                        }
+                    }
+                }
+
+                out.println("NAMEACCEPTED");
+                cs.writers.add(out);
+
+                while (true) {
+                    String input = in.readLine();
+                    if (input == null) {
+                        return;
+                    }
+                    for (PrintWriter writer : cs.writers) {
+                        writer.println("MESSAGE " + name + ": " + input);
+                    }
+                }
+            } catch (IOException e) {
+                System.out.println(e);
+            } finally {
+                
+            	if (name != null) {
+                    cs.names.remove(name);
+                }
+                if (out != null) {
+                    cs.writers.remove(out);
+                }
+                try {
+                    socket.close();
+                } catch (IOException e) {
+                }
+            }
+        }
+    }
+
